Update UI and change some logic for admin use case

+++ src/main/java/com/nhom8/camera/controller/admin/UserManageController.java
@@ -7,8 +7,10 @@ import com.nhom8.camera.model.request.UpdateUserAdminRequest;
 import com.nhom8.camera.model.response.UserResponse;
 import com.nhom8.camera.repository.RoleRepository;
 import com.nhom8.camera.repository.UserRepository;
+import com.nhom8.camera.security.CustomUserDetails;
 import com.nhom8.camera.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.ModelMap;
 import org.springframework.validation.BindingResult;
@@ -36,10 +38,10 @@ public class UserManageController {
     }
 
     @GetMapping("/list-user")
-    public ModelAndView listUser(@RequestParam(value = "page", defaultValue = "1") int page) {
+    public ModelAndView listUser(@RequestParam(value = "page", defaultValue = "1") int page, @AuthenticationPrincipal CustomUserDetails userDetails) {
         ModelAndView mav = new ModelAndView("/admin/user-table");
         int offset = (page-1)*limit;
-        List<User> users = userService.getUserAndSort(limit, offset);
+        List<User> users = userService.getUserAndSort(userDetails.getUsername(), limit, offset);
         long totalPage = userRepository.count()/limit + 1;
         mav.addObject("totalPage", totalPage);
         mav.addObject("users", users);
@@ -64,40 +66,44 @@ public class UserManageController {
     }
 
     @GetMapping("/user/{id}")
-    public ModelAndView updateUserGet(@PathVariable(value = "id") final Long id, @ModelAttribute(value = "userRequest") UpdateUserAdminRequest updateUserAdminRequest, @RequestParam(value = "page", defaultValue = "1") int page) {
-        ModelAndView mav = new ModelAndView("/admin/user-update");
-        List<Role> roles = roleRepository.findAll();
-        UserResponse user = userService.findUserById(id);
-        mav.addObject("roles", roles);
-        if (user !=  null) {
-            mav.addObject("user", user);
-            return mav;
+    public String updateUserGet(@PathVariable(value = "id") final Long id, @ModelAttribute(value = "userRequest") UpdateUserAdminRequest updateUserAdminRequest, @RequestParam(value = "page", defaultValue = "1") int page, ModelMap modelMap, @AuthenticationPrincipal CustomUserDetails userDetails) {
+        if (!id.equals(userDetails.getId())) {
+            List<Role> roles = roleRepository.findAll();
+            UserResponse user = userService.findUserById(id);
+            modelMap.addAttribute("roles", roles);
+            modelMap.addAttribute("user", user);
+            return "/admin/user-update";
         }
-        return listUser(page);
+        return "redirect:/admin/error";
     }
 
     @PostMapping("/user/{id}")
-    public String updateUserPost(@PathVariable(value = "id") final Long id, @Valid @ModelAttribute(value = "userRequest") UpdateUserAdminRequest updateUserAdminRequest, BindingResult result, ModelMap modelMap, @RequestParam(value = "page", defaultValue = "1") int page) {
-        User user = userRepository.findUserById(id);
-        String email = updateUserAdminRequest.getEmail();
-        if(!email.equals(user.getEmail()) && !userService.emailValid(email)) {
-            result.addError(new FieldError("updateUserAdminRequest", "email", "Email already exist or not same old email"));
+    public String updateUserPost(@PathVariable(value = "id") final Long id, @Valid @ModelAttribute(value = "userRequest") UpdateUserAdminRequest updateUserAdminRequest, BindingResult result, ModelMap modelMap, @AuthenticationPrincipal CustomUserDetails userDetails) {
+        if (!id.equals(userDetails.getId())) {
+            User user = userRepository.findUserById(id);
+            String email = updateUserAdminRequest.getEmail();
+            if(!email.equals(user.getEmail()) && !userService.emailValid(email)) {
+                result.addError(new FieldError("updateUserAdminRequest", "email", "Email already exist or not same old email"));
+            }
+            if(result.hasErrors()) {
+                List<Role> roles = roleRepository.findAll();
+                UserResponse userResponse = userService.findUserById(id);
+                modelMap.addAttribute("roles", roles);
+                modelMap.addAttribute("user", userResponse);
+                return "/admin/user-update";
+            }
+            userService.updateUserAdmin(user, updateUserAdminRequest, id);
+            return "redirect:/admin/list-user";
         }
-        if(result.hasErrors()) {
-//            ModelAndView mav = new ModelAndView("/admin/user-update");
-            List<Role> roles = roleRepository.findAll();
-            UserResponse userResponse = userService.findUserById(id);
-            modelMap.addAttribute("roles", roles);
-            modelMap.addAttribute("user", userResponse);
-            return "/admin/user-update";
-        }
-        userService.updateUserAdmin(user, updateUserAdminRequest, id);
-        return "redirect:/admin/list-user";
+        return "redirect:/admin/error";
     }
 
     @GetMapping("/user-delete/{id}")
-    public String deleteUser(@PathVariable Long id) {
-        userService.deleteUserAdmin(id);
-        return "redirect:/admin/list-user";
+    public String deleteUser(@PathVariable Long id, @AuthenticationPrincipal CustomUserDetails userDetails) {
+        if (!id.equals(userDetails.getId())) {
+            userService.deleteUserAdmin(id);
+            return "redirect:/admin/list-user";
+        }
+        return "redirect:/admin/error";
     }
 }
